requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
django-filter==25.1
drf-spectacular==0.28.0
Pillow==11.2.1
django-cors-headers==4.7.0
python-dotenv==1.0.1
web3==6.15.1
pycryptodome==3.20.0
PyYAML==6.0.2
gunicorn==22.0.0
whitenoise==6.6.0